create linux and darwing flavors for pie test
